Add timing to build, cache layouts.
Adds metalsmith-timer to each step of the build. This can probably be
refactored for DRY, but works for now. Pass the `cache` option to
layouts which gets passed to pug. The layouts plugin does not compile
templates, this allows pug to cache the compiled version itself. For me
the pug stage of the build goes from 70s to <1s.

Autotoc is now the slowest stage at 15s. Much more usable this way.