Remove application-logging page with dated, incorrect information.
I imagine something like this would be useful but it should probably be generated from code artifacts and live closer to the actual code. I don't see the team as having the throughput to update and keep this updated in the meantime though - if I'm wrong someone should step up and offer to maintain the page.