Update running tests documentation to reflect newer source of information.
Most of the information available in that file is dated and incorrect, most of what was incorret is correct in `run_tests.sh --help`, most of what was correct was duplicated in that help command, and the remaining difference I think I covered with https://github.com/galaxyproject/galaxy/pull/3715.

What documentation we do have is insufficient, but having another source of this information that is incorrect isn't helping anything. Hopefully we can continue to improve run_tests.sh and then ultimately spin it off to a Markdown document that ships with the code and appears in the Galaxy code docs so that we can make sure changes to Galaxy testing framework also change the documentation atomically.