virtualenv is not working anymore
I have added some modifications, but please have a closer look as I only tested it on 1 server so far. The python virtualenv from the wget is from 2011 and was causing major issues on our server. Using the virtualenv from pip seems to have worked so far. Further testing in progress.