Collaborate: non-GCC projects, additional pointers
Dave -- this looks excellent. I added some additional color around why
we moved the collaboration to the end, and also provide pointers to
non-GCC projects. I moved that up the list a bit so folks definitely
know this is not Galaxy specific and we welcome anyone interested in any
open source bioinformatics project.