Adapt eLabFTW blog post to deployment based on file source templates
On public instances, it makes sense to deploy the eLabFTW file source using file source templates (and that is how it is deployed on usegalaxy.eu) because there is no public eLabFTW instance. Therefore, adapt the blog post to this kind of deployment.

For private Galaxy instances, the contrary might make better sense, since all researchers using Galaxy are likely to be part of the same lab and use the same instance.